Fix CI: real cross-version failures under Python 3.10/3.11
The pushed CI run failed Type check on the 3.10 and 3.11 legs (3.12/3.13 and the
wheel smoke test were green). Root cause: mypy infers its target Python version
from the interpreter it runs under when none is pinned, so my local checks
(Python 3.13) never actually type-checked the 3.10/3.11 code paths this package
claims to support. Reproduced with real 3.10/3.11 interpreters via uv (matching
CI exactly), not guessed at.

Fixes, verified clean under 3.10/3.11/3.12/3.13:
- _compat.py: typing.Self backport for 3.10 (typing_extensions), explicit
  StrEnum __str__/__format__ methods instead of assigning str's dunders (mypy
  types str.__format__ as unbound and rejects the assignment).
- Bare `np.ndarray` annotations in 5 test files replaced with the project's
  own FloatArray/IntArray/AnyArray aliases (numpy's stub needs an explicit
  type arg under an older mypy target).
- Several `.all(axis=1)` boolean-mask expressions explicitly cast to
  BoolArray, where mypy's overload resolution returns a wider union under
  3.10/3.11.

Also fixed, found by actually running the suite under 3.10 (not just mypy):
- causal.py: _linear_r2 could feed non-finite lstsq coefficients into a
  matmul, propagating -inf/nan into a detector's R^2. A real bug, not a
  warning artifact: now checked and short-circuited to 0.0.
- Four sites (shapes.py x2, twosample.py, dynamics_model.py) construct a
  matmul on already-finite-validated input but still triggered spurious
  RuntimeWarnings from certain BLAS backends on Python 3.10 (reproduced with
  uv's own managed interpreter, not just a local Homebrew build) inside
  sklearn's GaussianMixture/Ridge and our own distance kernels. Scoped
  np.errstate around each, since the surrounding code already guards the one
  thing that could actually be wrong (clamped to 0.0, or checked finite).

Also commits uv.lock: CI's setup-uv step was warning the cache can never
invalidate without one.

Verified the full matrix (pytest + mypy + ruff) green under 3.10, 3.11, 3.12,
and 3.13 via uv, plus a clean-venv install and scan of the built wheel.

+++ tests/test_lag_conventions.py
@@ -216,7 +216,7 @@ def test_best_lag_refuses_a_boundary_peak() -> None:
     arbitrary, which would not exercise the boundary rule.
     """
     rng = np.random.default_rng(3)
-    smooth = np.convolve(rng.normal(size=240), np.ones(9) / 9, mode="valid")[:200]
+    smooth = np.convolve(rng.normal(size=240), np.ones(9) / 9, mode="valid")[:200].astype(np.float64)
     assert _best_lag(smooth, np.roll(smooth, 9), 4) is None  # true offset far outside ±4
